requirements.txt
PySide6-Essentials
Pillow
